+++ server/src/main/java/com/omigost/server/rest/SnsHandlerController.java
@@ -2,13 +2,16 @@ package com.omigost.server.rest;
 
 import com.amazonaws.services.budgets.model.Budget;
 import com.omigost.server.aws.BudgetService;
+import com.omigost.server.exception.BudgetNotFound;
 import com.omigost.server.notification.MainNotificationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/alerts")
 public class SnsHandlerController {
 
     @Autowired
@@ -17,7 +20,7 @@ public class SnsHandlerController {
     @Autowired
     private MainNotificationService notifications; // TODO should notifications be user-based? budget-based? where to store config and where to consider it in code
 
-    @PostMapping("/alerts/trigger")
+    @PostMapping("/trigger")
     public void handleAlertTrigger(@RequestParam String budgetName) {
         Budget budget = budgets.getBudgetByName(budgetName);
         if (budget == null) throw new BudgetNotFound();
